Git-a-fy the repo
Get rid of old versions, flatten tree, .ignore artifact files.
